Refactor out brace whitespace tests
